+++ mxf/mxf_app.h
@@ -69,12 +69,12 @@ int mxf_app_load_extensions(MXFDataModel *dataModel);
 
 int mxf_app_is_app_mxf(MXFHeaderMetadata *headerMetadata);
 int mxf_app_get_info(MXFHeaderMetadata *headerMetadata, ArchiveMXFInfo *info);
-int mxf_app_get_pse_failures(MXFHeaderMetadata *headerMetadata, PSEFailure **failures, long *numFailures);
-int mxf_app_get_vtr_errors(MXFHeaderMetadata *headerMetadata, VTRErrorAtPos **errors, long *numErrors);
+int mxf_app_get_pse_failures(MXFHeaderMetadata *headerMetadata, PSEFailure **failures, size_t *numFailures);
+int mxf_app_get_vtr_errors(MXFHeaderMetadata *headerMetadata, VTRErrorAtPos **errors, size_t *numErrors);
 int mxf_app_get_digibeta_dropouts(MXFHeaderMetadata *headerMetadata, DigiBetaDropout **digibetaDropouts,
-                                  long *numDigiBetaDropouts);
+                                  size_t *numDigiBetaDropouts);
 int mxf_app_get_timecode_breaks(MXFHeaderMetadata *headerMetadata, TimecodeBreak **timecodeBreaks,
-                                long *numTimecodeBreaks);
+                                size_t *numTimecodeBreaks);
 
 
 /* returns 1 if footer headermetadata was read, return 2 if none is present (*headerMetadata is NULL) */
